Created DatabaseDownloadTrait.
