+++ src/Robo/Plugin/Commands/DevelopmentModeBaseCommands.php
@@ -2,7 +2,7 @@
 
 namespace ChqRobo\Robo\Plugin\Commands;
 
-use AsyncAws\S3\S3Client;
+use ChqRobo\Robo\Plugin\Traits\DatabaseDownloadTrait;
 use ChqRobo\Robo\Plugin\Traits\SitesConfigTrait;
 use DrupalFinder\DrupalFinder;
 use Robo\Exception\TaskException;
@@ -18,9 +18,8 @@ use Symfony\Component\Yaml\Yaml;
  */
 class DevelopmentModeBaseCommands extends Tasks
 {
+    use DatabaseDownloadTrait;
     use SitesConfigTrait;
-
-    protected const S3_DEFAULT_REGION = 'us-east-1';
 
     /**
      * Drupal root directory.
@@ -49,59 +48,6 @@ class DevelopmentModeBaseCommands extends Tasks
         $drupalFinder->locateRoot(getcwd());
         $this->drupalRoot = $drupalFinder->getDrupalRoot();
         $this->devServicesPath = "$this->drupalRoot/sites/fe.development.services.yml";
-    }
-
-    /**
-     * Download the latest database dump for the site.
-     *
-     * @param string $siteName
-     *   The site name.
-     *
-     * @aliases dbdl
-     *
-     * @return string|\Robo\Result
-     *   The path of the last downloaded database.
-     *
-     * @throws \Robo\Exception\TaskException
-     */
-    public function databaseDownload(string $siteName = 'default')
-    {
-        $this->io()->title('database download.');
-
-        $awsConfigDirPath = getenv('HOME') . '/.aws';
-        $awsConfigFilePath = "$awsConfigDirPath/credentials";
-        if (!is_dir($awsConfigDirPath) || !file_exists($awsConfigFilePath)) {
-            $result = $this->configureAwsCredentials($awsConfigDirPath, $awsConfigFilePath);
-            if ($result->wasCancelled()) {
-                return Result::cancelled();
-            }
-        }
-
-        $s3 = new S3Client([
-            'region' => $this->s3RegionForSite($siteName),
-        ]);
-        $objects = $s3->listObjectsV2($this->s3BucketRequestConfig($siteName));
-        $objects = iterator_to_array($objects);
-        if (empty($objects)) {
-            throw new TaskException($this, "No database dumps found for '$siteName'.");
-        }
-        // Ensure objects are sorted by last modified date.
-        usort($objects, fn($a, $b) => $a->getLastModified()->getTimestamp() <=> $b->getLastModified()->getTimestamp());
-        $latestDatabaseDump = array_pop($objects);
-        $dbFilename = $latestDatabaseDump->getKey();
-
-        if (file_exists($dbFilename)) {
-            $this->say("Skipping download. Latest database dump file exists >>> $dbFilename");
-        } else {
-            $result = $s3->GetObject([
-                'Bucket' => $this->s3BucketForSite($siteName),
-                'Key' => $dbFilename,
-            ]);
-            $fp = fopen($dbFilename, 'wb');
-            stream_copy_to_stream($result->getBody()->getContentAsResource(), $fp);
-            $this->say("Database dump file downloaded >>> $dbFilename");
-        }
-        return $dbFilename;
     }
 
     /**
@@ -262,105 +208,6 @@ class DevelopmentModeBaseCommands extends Tasks
         $result = $this->databaseRefreshLando($siteName);
         $result = $this->drupalLoginLink($siteName);
         return $result;
-    }
-
-    /**
-     * Configure AWS credentials.
-     *
-     * @param string $awsConfigDirPath
-     *   Path to the AWS configuration directory.
-     * @param string $awsConfigFilePath
-     *   Path to the AWS configuration file.
-     *
-     * @return \Robo\Result
-     */
-    protected function configureAwsCredentials(string $awsConfigDirPath, string $awsConfigFilePath)
-    {
-        $yes = $this->io()->confirm('AWS S3 credentials not detected. Do you wish to configure them?');
-        if (!$yes) {
-            return Result::cancelled();
-        }
-
-        if (!is_dir($awsConfigDirPath)) {
-            $this->_mkdir($awsConfigDirPath);
-        }
-
-        if (!file_exists($awsConfigFilePath)) {
-            $this->_touch($awsConfigFilePath);
-        }
-
-        $awsKeyId = $this->ask("AWS Access Key ID:");
-        $awsSecretKey = $this->askHidden("AWS Secret Access Key:");
-        return $this->taskWriteToFile($awsConfigFilePath)
-            ->line('[default]')
-            ->line("aws_access_key_id = $awsKeyId")
-            ->line("aws_secret_access_key = $awsSecretKey")
-            ->run();
-    }
-
-    /**
-     * Build S3 request configuration from sites config.
-     *
-     * @param string $siteName
-     *   The site name.
-     *
-     * @return array
-     *   An S3 request object configuration array.
-     */
-    protected function s3BucketRequestConfig(string $siteName): array
-    {
-        $s3ConfigArray = ['Bucket' => $this->s3BucketForSite($siteName)];
-        try {
-            $s3KeyPrefix = $this->getConfig('database_s3_key_prefix_string', $siteName);
-            $this->say("'$siteName' S3 Key prefix: '$s3KeyPrefix'");
-            $s3ConfigArray['Prefix'] = $s3KeyPrefix;
-        } catch (TaskException $e) {
-            $this->say("No S3 Key prefix found for $siteName.");
-        }
-        return $s3ConfigArray;
-    }
-
-    /**
-     * Get S3 Bucket for site.
-     *
-     * @param string $siteName
-     *   The site name.
-     *
-     * @return string
-     *   An S3 bucket.
-     *
-     * @throws \Robo\Exception\TaskException
-     */
-    protected function s3BucketForSite(string $siteName): string
-    {
-        if (!$bucket = $this->getConfig('database_s3_bucket', $siteName)) {
-            throw new TaskException($this, "database_s3_bucket value not set for '$siteName'.");
-        }
-        $this->say("'$siteName' S3 bucket: $bucket");
-        return $bucket;
-    }
-
-    /**
-     * Get S3 region for site.
-     *
-     * @param string $siteName
-     *   The site name.
-     *
-     * @return string
-     *   An S3 region.
-     */
-    protected function s3RegionForSite(string $siteName): string
-    {
-        try {
-            $region = $this->getConfig('database_s3_region', $siteName);
-            $this->say("'$siteName' database_s3_region set to $region.");
-        } catch (TaskException $e) {
-            // Set default region if one is not set.
-            $defaultRegion = self::S3_DEFAULT_REGION;
-            $this->say("'$siteName' database_s3_region not set. Defaulting to $defaultRegion.");
-            $region = $defaultRegion;
-        }
-        return $region;
     }
 
     /**
